requirements.txt
numpy
openai-whisper
sounddevice
rich
langchain
langchain-community
pyttsx3
